separate roundmaker and outbarround
