+++ src/routes/round/index.tsx
@@ -14,9 +14,9 @@ const Round = () => {
   const settingContext = TournamentStore();
 
   //// testar att skapa en ny array från players-context //////
-  console.log("loggar constext vid refresh", settingContext.playerList);
+
   const playerArray: any = [];
-  console.log("players från round context", settingContext.playerList);
+
   const getNameAndId = () => {
     settingContext.tournament.players.forEach((e) => {
       playerArray.push({ name: e.name, id: e.id });
